pyproject.toml
[project]
name = "Gemini-Image-Tool"
version = "0.2.0"
description = ""
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "flet==0.80.0",
    "google-genai>=1.52.0",
    "pillow>=11.3.0",
    "requests>=2.32.5",
    "appdirs>=1.4.3"
]
authors = [
    { name = "joycai", email = "joycai@joycai.cn" }
]

[tool.flet]
org = ""
product = "Gemini-Image-Tool"
company = "Joycai"
copyright = "Copyright (C) 2025 by Joycai"
app.path = "."
app.module = "flet_app"

[tool.uv]
dev-dependencies = [
    "flet[all]==0.80.0",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
flet = {extras = ["all"], version = "0.28.3"}

[tool.flet.app.boot_screen]
show = false
message = "Preparing the app for its first launch…"

[tool.flet.app.startup_screen]
show = false
message = "Starting up the app…"